small review changes
