+++ metals/src/main/scala/scala/meta/internal/rename/RenameProvider.scala
@@ -244,7 +244,7 @@ final class RenameProvider(
                     List(implReferences, currentReferences, companionRefs)
                   )
                   .map(
-                    _.reduce(_ ++ _) ++ definitionLocation
+                    _.flatten ++ definitionLocation
                   )
               Future
                 .sequence(allReferences)
