+++ mtags/src/main/scala-3/scala/meta/internal/pc/PcSymbolSearch.scala
@@ -47,9 +47,8 @@ trait PcSymbolSearch:
       Interactive.pathTo(sel, pos.span) ::: rawPath
     case _ => rawPath
 
-  lazy val soughtSymbols: Option[(Set[Symbol], SourcePosition)] = soughtSymbols(
-    path
-  )
+  lazy val soughtSymbols: Option[(Set[Symbol], SourcePosition)] =
+    soughtSymbols(path)
 
   def soughtSymbols(path: List[Tree]): Option[(Set[Symbol], SourcePosition)] =
     val sought = path match
@@ -173,64 +172,6 @@ trait PcSymbolSearch:
       case _ => sought
 
   end soughtSymbols
-
-  // First identify the symbol we are at, comments identify @@ as current cursor position
-  def symbolAlternatives(sym: Symbol) =
-    def member(parent: Symbol) = parent.info.member(sym.name).symbol
-    def primaryConstructorTypeParam(owner: Symbol) =
-      for
-        typeParams <- owner.primaryConstructor.paramSymss.headOption
-        param <- typeParams.find(_.name == sym.name)
-        if (param.isType)
-      yield param
-    def additionalForEnumTypeParam(enumClass: Symbol) =
-      if enumClass.is(Flags.Enum) then
-        val enumOwner =
-          if enumClass.is(Flags.Case)
-          then
-            // we check that the type parameter is the one from enum class
-            // and not an enum case type parameter with the same name
-            Option.when(member(enumClass).is(Flags.Synthetic))(
-              enumClass.maybeOwner.companionClass
-            )
-          else Some(enumClass)
-        enumOwner.toSet.flatMap { enumOwner =>
-          val symsInEnumCases = enumOwner.children.toSet.flatMap(enumCase =>
-            if member(enumCase).is(Flags.Synthetic)
-            then primaryConstructorTypeParam(enumCase)
-            else None
-          )
-          val symsInEnumOwner =
-            primaryConstructorTypeParam(enumOwner).toSet + member(enumOwner)
-          symsInEnumCases ++ symsInEnumOwner
-        }
-      else Set.empty
-    val all =
-      if sym.is(Flags.ModuleClass) then
-        Set(sym, sym.companionModule, sym.companionModule.companion)
-      else if sym.isClass then
-        Set(sym, sym.companionModule, sym.companion.moduleClass)
-      else if sym.is(Flags.Module) then
-        Set(sym, sym.companionClass, sym.moduleClass)
-      else if sym.isTerm && (sym.owner.isClass || sym.owner.isConstructor)
-      then
-        val info =
-          if sym.owner.isClass then sym.owner.info else sym.owner.owner.info
-        Set(
-          sym,
-          info.member(sym.asTerm.name.setterName).symbol,
-          info.member(sym.asTerm.name.getterName).symbol,
-        ) ++ sym.allOverriddenSymbols.toSet
-      // type used in primary constructor will not match the one used in the class
-      else if sym.isTypeParam && sym.owner.isPrimaryConstructor then
-        Set(sym, member(sym.maybeOwner.maybeOwner))
-          ++ additionalForEnumTypeParam(sym.maybeOwner.maybeOwner)
-      else if sym.isTypeParam then
-        primaryConstructorTypeParam(sym.maybeOwner).toSet
-          ++ additionalForEnumTypeParam(sym.maybeOwner) + sym
-      else Set(sym)
-    all.filter(s => s != NoSymbol && !s.isError)
-  end symbolAlternatives
 
   private def seekInExtensionParameters() =
     def collectParams(
